Moved the FeatureCollectionDataAdapter Benchmark to a new benchmark module
