+++ geowave-benchmark/src/test/resources/log4j.properties
@@ -1,7 +1,7 @@
 log4j.rootLogger=ERROR,stdout,fout
 
-log4j.logger.mil.nga.giat.geowave.vector.test=INFO
-log4j.logger.mil.nga.giat.geowave.vector.util=INFO
+log4j.logger.mil.nga.giat.geowave.benchmark=INFO
+log4j.logger.mil.nga.giat.geowave.benchmark.util=INFO
 log4j.logger.mil.nga.giat.geowave.accumulo.BasicAccumuloOperations=WARN
 
 log4j.appender.stdout=org.apache.log4j.ConsoleAppender
